credit_service/requirements.txt
flask==2.3.2
flask-sqlalchemy==3.0.5
python-dotenv==1.0.0
PyJWT==2.7.0
requests==2.31.0
pymysql==1.0.3